Changed timezone hack for compile-time unix time.
compile_time.h is now verbatim from stackoverflow (except for a direct
link to the answer where it was found).

firmware.ino now does the timezone offsetting when initializing the
constant, including a rough estimate of whether DST is in effect.

Also moved the UNIX_OFFSET subtraction to constant initialization so it
doesn't have to be done at runtime.