Add code to detect key presses.
The HT16K33 INT pin is configured active high, and is connected to
digital 3 on the Uno.  Rising-edge interrupt is used on Arduino to
detect when the HT16K33 brings its INT pin high.

Unfortunately, contrary to a statement in the datasheet, the HT16K33
will re-assert the INT pin if the key continues to be held down more
than 2 scan cycles after reading key data.  So there is a polling logic,
which starts when the INT pin first goes high, and ends when a poll
reports no keys are pressed.

Also added the 5 LEDs (for lighted buttons) on keypad to be used to
display the least-significant 5 bits of the seconds field, as a proof
that they work and are individually addressable.