Hook up input via keypad.
The rules are fairly complicated, and extrapolate from the movie clips
and fan wiki:

First, you press the color-lighted button corresponding to which color
display you want to set.  This clears that display, and puts you into
"entry" mode for that display.  To further indicate this (not how it
seems to be in the movie) it turns off the color button for the color
you are entering.  It also turns off the second-from-bottom, white
button.  Then you enter the date/time as MMDDYYYYHHMM
(month-day-year-hour-minute, 24-hour time), and press the bottom left
button, which acts as "enter".  This commits the entry, displays it on
the proper display, and turns on the button lights for the display that
was just entered and the second-from-bottom, white button (as seen in
movie clip).

As an additional non-movie extension, the bottom button is lit when the
display is showing what the firmware believes to be the current time
(right now hooked to the red display, because I haven't built any other
colors).

Additionally, while in "entry" mode, the second-from-the-bottom button
acts as a sort of backspace, erasing the last entered digit, and if
backspaced past the beginning of "entry" mode will reset it back to
"current" time mode.  Note that backspace doesn't really work 100%
because of the limited number of bits I gave the various fields: the
2-digit fields can't hold 0-99, only 0-15 for month, 0-31 for day or
hour, and 0-64 for minute.  If you accidentally entered a tens value
outside that range, and then immediately backspace it, the value may not
go back to zero as intended.  It should work if you enter a ones place
and then do 2 backspaces...