Swap green and yellow displays in code.
Managed to confuse which order they were in in the movie.  The keypad,
as well as common sense, goes red yellow green, but the displays are red
green yellow.  Not horribly important as long as the addresses are
right, but may as well process them in the order they're displayed.