Add ability to set current time from keypad.
Previously the current time was set either from the compile time, or
from a unix time input via serial.

Now, if the row which shows the current time is in "override" mode, and
no row is in "entry" mode, then hitting the "backspace" (left column
2nd-from-bottom) key will set the current time to the override time for
the row which shows the current time, and take that row out of
"override" mode (which means it will revert to showing the current time,
which for the next minute will match the prior override time).